% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformIntensities.R
\name{convertNAToStrings}
\alias{convertNAToStrings}
\title{Convert columns to characters and NA to strings}
\usage{
convertNAToStrings(outputTable)
}
\description{
This is needed for the parquet conversion
}
\keyword{internal}
