% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformIntensities.R
\name{transformIntensities}
\alias{transformIntensities}
\title{Transform intensities}
\usage{
transformIntensities(
  intensitiesTable,
  featuresMetadataTable,
  featureColname,
  replicateColname,
  normMethod
)
}
\arguments{
\item{intensitiesTable}{data.frame. Mass Dynamics intensities table.}

\item{featuresMetadataTable}{data.frame Mass Dynamics metadata table.}

\item{featureColname}{str Name of feature column name in `intensitiesTable` and `featuresMetadataTable`.}

\item{replicateColname}{str Name of replicate column name in `intensitiesTable`.}

\item{normMethod}{str Normalisation method to pass to `limma::normalizeBetweenArrays`.}
}
\description{
Transform intensities
}
