% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformIntensities.R
\name{transformIntensities}
\alias{transformIntensities}
\title{Transform intensities}
\usage{
transformIntensities(
  intensities,
  metadata,
  featureColname = "GroupId",
  replicateColname = "replicate",
  normMethod
)
}
\arguments{
\item{intensities}{data.frame. Mass Dynamics intensities table.}

\item{metadata}{data.frame Mass Dynamics metadata table.}

\item{featureColname}{str Name of feature column name in `intensities` and `metadata`.}

\item{replicateColname}{str Name of replicate column name in `intensities`.}

\item{normMethod}{str Normalisation method to pass to `limma::normalizeBetweenArrays`.}
}
\description{
Transform intensities
}
